+++ packages/shadow-objects/src/shae-ent.ts
@@ -1,4 +1,4 @@
-import {ShaeEntElement} from './elements/ShaeEntElement.js';
 import {SHAE_ENT} from './elements/constants.js';
+import {ShaeEntElement} from './elements/ShaeEntElement.js';
 
 customElements.define(SHAE_ENT, ShaeEntElement);