+++ packages/shadow-objects/src/shae-worker.ts
@@ -1,4 +1,4 @@
-import {ShaeWorkerElement} from './elements/ShaeWorkerElement.js';
 import {SHAE_WORKER} from './elements/constants.js';
+import {ShaeWorkerElement} from './elements/ShaeWorkerElement.js';
 
 customElements.define(SHAE_WORKER, ShaeWorkerElement);